solve validation problem

+++ Backend/artwork-space-master/src/main/java/com/system/artworkspace/collection/CollectionController.java
@@ -2,10 +2,16 @@ package com.system.artworkspace.collection;
 
 import com.system.artworkspace.artwork.ArtworkDto;
 import com.system.artworkspace.artwork.ArtworkMapper;
+import com.system.artworkspace.exceptions.ExceptionHelper;
+import com.system.artworkspace.exceptions.ValidationException;
+import javax.validation.Valid;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -30,7 +36,11 @@ public class CollectionController {
         return collectionService.getAllCollections().stream().map(x-> CollectionMapper.INSTANCE.collectionToCollectionDto(x)).collect(Collectors.toList());
     }
     @PostMapping
-    public CollectionDto createCollection(@RequestBody CollectionDto collection) {
+    public CollectionDto createCollection(@RequestBody @Valid CollectionDto collection, BindingResult bindingResult) {
+        if (bindingResult.hasErrors()) {
+            String message = ExceptionHelper.formErrorMessage(bindingResult);
+            throw new ValidationException(message);
+        }
         logger.info("Creating a collection with ID: {}", collection.getId());
         CollectionDto createdCollection = CollectionMapper.INSTANCE.collectionToCollectionDto(collectionService.createCollection(CollectionMapper.INSTANCE.collectionDtoToCollection(collection)));
         logger.info("Collection created with ID: {}", createdCollection.getId());
@@ -38,7 +48,11 @@ public class CollectionController {
     }
 
     @PostMapping("/{collectionId}/addArtwork")
-    public void addToCollection(@PathVariable Long collectionId, @RequestBody ArtworkDto artwork) {
+    public void addToCollection(@PathVariable Long collectionId, @RequestBody @Valid ArtworkDto artwork, BindingResult bindingResult) {
+        if (bindingResult.hasErrors()) {
+            String message = ExceptionHelper.formErrorMessage(bindingResult);
+            throw new ValidationException(message);
+        }
         logger.info("Adding artwork with ID {} to collection with ID: {}", artwork.getId(), collectionId);
         collectionService.addToCollection(collectionId, ArtworkMapper.INSTANCE.artworkDtoToArtwork(artwork));
         logger.info("Artwork added to collection with ID: {}", collectionId);
@@ -53,12 +67,21 @@ public class CollectionController {
 
     //TODO think about using artwork id
     @DeleteMapping("/{collectionId}/removeArtwork")
-    public void deleteFromCollection(@PathVariable Long collectionId, @RequestBody ArtworkDto artwork) {
+    public void deleteFromCollection(@PathVariable Long collectionId, @RequestBody @Valid ArtworkDto artwork, BindingResult bindingResult) {
+        if (bindingResult.hasErrors()) {
+            String message = ExceptionHelper.formErrorMessage(bindingResult);
+            throw new ValidationException(message);
+        }
         logger.info("Removing artwork with ID {} from collection with ID: {}", artwork.getId(), collectionId);
         collectionService.deleteFromCollection(collectionId, ArtworkMapper.INSTANCE.artworkDtoToArtwork(artwork));
         logger.info("Artwork removed from collection with ID: {}", collectionId);
     }
-
+    @ExceptionHandler(Exception.class)
+    public ResponseEntity<String> handleException(Exception e) {
+        String errorMessage = "ERROR: " + e.getMessage();
+        logger.error(errorMessage);
+        return new ResponseEntity<>(errorMessage, HttpStatus.INTERNAL_SERVER_ERROR);
+    }
     //Todo think about only id
     @PutMapping("/{collectionId}/editName")
     public void editName(@PathVariable Long collectionId, @RequestParam String name) {
